django_restful/requirements.txt
###################
# pipenv run pip freeze > requirements.txt
###################

amqp==2.6.1
aniso8601==7.0.0
arrow==1.2.3
async-timeout==4.0.2
autopep8==1.7.0
billiard==3.6.3.0
celery==4.4.2
Cerberus==1.3.4
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==2.2.16
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
dnspython==1.16.0
eventlet==0.29.0
flake8==5.0.4
flower==0.9.5
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==0.4.17
humanize==3.1.0
idna==3.4
importlib-metadata==4.2.0
kombu==4.6.11
mccabe==0.7.0
mimesis==4.1.2
monotonic==1.5
mysqlclient==2.1.1
packaging==21.3
prometheus-client==0.8.0
promise==2.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
shortuuid==1.0.9
singledispatch==3.4.0.3
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tornado==6.0.4
typing_extensions==4.4.0
Unidecode==1.1.1
urllib3==1.26.12
vine==1.3.0
wrapt==1.14.1
zipp==3.9.0
